Move Where shape inference into the binary submodule
Move the Where operator, plus its test, from ops.rs into ops/binary.rs.
Where is not a binary operator, but this matches how it is grouped in
src/ops/binary_elementwise.rs.
